requirements.txt
python-vlc
ollama
openpyxl
